(f)web:theme: correctly point to the local avatar path
link: drop 90cd06ea8f4f6880a5c614269b1ff86c604830b8 11f1a43f347a44d2daf3a73ec9656414e38c8c13 This is a fix for both commits, as they fu*k were both wrong.

We put the avatar images in /source/images because the files that existed in public before git doesn't keep track of him, '/public' is ignored by '.gitignore' because vercel will rebuild hexo to produce a new '/public' in vercel server.
And after that we create a separate '/external_pu/custom/' directory and put the files in this directory, but the relative path filled in the redefine theme config yml redefine thinks '/' is under '/source', so We simply put the files in /source/images.

Signed-off-by: Hou Qiyao <shinka830@outlook.jp>